Create main.yml
Caches GMP repos to GitHub, using
https://github.com/mozillazg/job-mirror-hg-repos.

There's primitive caching between runs so that a full 'hg clone' is only
done if the commit ID of the tip has changed since the last run, or a
full clone is done once a week in case there's activity on a branch
(which seems unlikely).

Older repos that seem dormant are disabled (but could be easily
re-enabled if needed).